refactor(frontend): move characters.ts to utils/
Shared game data used across features (game, chat, local-solo) belongs in utils, not at the src root.
