+++ srcs/frontend/src/features/game/components/GoalFlash.tsx
@@ -1,6 +1,6 @@
 import { useMemo } from 'react'
 import { useTranslation } from 'react-i18next'
-import { flagSrc } from '../../../characters'
+import { flagSrc } from '../../../utils/characters'
 import { GOAL_FLASH_DURATION_MS } from '../engine/constants'
 
 interface GoalFlashProps {
