+++ srcs/frontend/src/features/local-solo/CarouselLocalPlayer1.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { FaPlay } from "react-icons/fa";
-import { CHARACTERS, faceSrc, flagSrc } from '../../characters';
+import { CHARACTERS, faceSrc, flagSrc } from '../../utils/characters';
 
 export const Carouselplayer1 = ({ onChange }: { onChange?: (char: string) => void }) => {
     const [activeIndex, setActiveIndex] = useState<number>(0);
