+++ srcs/frontend/src/utils/characters.ts
